Add preview subsystem
It works, but it's only minimally tested

Beginnings of the preview subsystem

It works but may be a little broken. I'm trying to hack together a way
of re-parenting the Xephyr window as a QMdiSubWindow but not having very
much luck. This commit is to save the current "working" state of the
preview class so that I can attempt to subclass some other kind of
window maybe. It's not looking likely.

Add preview feature in alpha state