Add LICENSE and remove vestigial script
get_prev_id.sh had been used to grab a window created by the application
that was used as a parent for Xephyr. I was not able to get Xephyr to
properly reparent to a QMdiSubWindow, so this command is unnecessary.
Now, we simply open the instance of Xephyr in its own window.